meson: Override dependencies and programs
This allows built programs to be used in the parent project. Unlike for
dependencies, Meson doesn't allow overriding program names in the .wrap
file. So, we have to override them in the source. This is the
recommended practice anyway, so we apply it to dependencies as well.

This also fixes the variable name for executables, which was previously
using only the first character of the name instead of the whole string.

Part-of: <https://gitlab.freedesktop.org/gstreamer/meson-ports/ffmpeg/-/merge_requests/141>