meson: Add an option to suppress warnings as a subproject
When building as a part of GStreamer, FFmpeg *always* emits some
compiler warnings that realistically we cannot fix. Upstream does not
seem to care, and they also do not accept patches when we submit them.

Add an option to suppress all warnings when running as a subproject.

Part-of: <https://gitlab.freedesktop.org/gstreamer/meson-ports/ffmpeg/-/merge_requests/162>